refactor(core): split core_api more more
