+++ core/src/lib.rs
@@ -177,9 +177,11 @@ pub mod mem_check;
 pub mod api {
     pub mod api_state;
     pub mod data_sources;
+    pub mod databases;
     pub mod datasets;
     pub mod projects;
     pub(crate) mod run_manager;
     pub mod runs;
     pub mod specifications;
+    pub mod tables;
 }
